refactor(core): extract parseInlineRubrics to shared method
- Add parseInlineRubrics function to evaluator-parser.ts
- Update yaml-parser.ts and jsonl-parser.ts to use shared function
- Eliminates ~50 lines of duplicated inline rubrics parsing logic
