+++ .changeset/shared-inline-rubrics.md
@@ -1,0 +1,5 @@
+---
+"@agentv/core": patch
+---
+
+Refactor: Extract inline rubrics parsing to shared `parseInlineRubrics` function
